[fix] 위치 설정 모달로 변경, 위치 설정 변경 가능하도록 수정
