+++ dutchiepay/src/app/_components/_community/_post/MeetingPlaceInput.jsx
@@ -1,4 +1,7 @@
-export default function MettingPlaceInput({ locationDescription }) {
+export default function MettingPlaceInput({
+  locationDescription,
+  setIsModalOpen,
+}) {
   return (
     <>
       <div className="flex items-center gap-[12px] mt-[24px] mb-[8px]">
@@ -8,11 +11,17 @@ export default function MettingPlaceInput({ locationDescription }) {
         <small className="text-sm text-gray--500">
           거래를 진행할 장소입니다.
         </small>
+        <button
+          className="text-sm text-blue-500 font-bold hover:underline"
+          onClick={() => setIsModalOpen(true)}
+        >
+          거래위치변경
+        </button>
       </div>
       <input
         className="w-[600px] border rounded-lg outline-none py-[8px] px-[12px] placeholder:text-sm"
         type="text"
-        value={`${locationDescription}`}
+        value={locationDescription || ''}
         placeholder="거래 장소"
         disabled={true}
       />
